[zephyr] Move HAL bodies into components/zephyr/hal.cpp + inline trivial dispatches
Per-platform follow-up to #15977.

Move out-of-line HAL bodies (arch_init with watchdog setup, arch_feed_wdt
with conditional wdt_feed, arch_restart) from components/zephyr/core.cpp
into a new components/zephyr/hal.cpp. core.cpp keeps the Mutex /
InterruptLock methods, random_bytes, get_mac_address_raw, and main().

Inline the trivial Zephyr-API one-liners directly in hal_zephyr.h.
The header pulls <zephyr/kernel.h> for the underlying primitives — that
header is already included by every Zephyr TU on this platform, so the
extra include cost is minimal:

  yield()                      -> ::k_yield()
  delay(ms)                    -> ::k_msleep(ms)
  micros()                     -> k_ticks_to_us_floor32(k_uptime_ticks())
  millis_64()                  -> k_uptime_get()
  millis()                     -> static_cast<uint32_t>(millis_64())
  delayMicroseconds(us)        -> ::k_usleep(us)
  arch_get_cpu_cycle_count()   -> k_cycle_get_32()
  arch_get_cpu_freq_hz()       -> sys_clock_hw_cycles_per_sec()
